Initial setup.
Initial setup for tools-repository-splitter-updater. This work is based on the
the NZQA git repository m11n-pipeline-eqa-code-updater, but has been
significantly modified in terms of variable names and some flows to reflect
generic splitting and updating of git repositories.

See the README.md for more details on how this tool works and its usage.